Initial commit of sequence number tracking in the producer

+++ clients/src/main/java/org/apache/kafka/common/record/MemoryLogBuffer.java
@@ -306,7 +306,7 @@ public class MemoryLogBuffer extends AbstractLogBuffer {
                                                  long baseOffset,
                                                  long pid,
                                                  short epoch,
-                                                 int baseSequence) {
+                                                 long baseSequence) {
         return new MemoryLogBufferBuilder(buffer, magic, compressionType, timestampType, baseOffset,
                 System.currentTimeMillis(), pid, epoch, baseSequence, buffer.capacity());
     }
