+++ clients/src/main/java/org/apache/kafka/common/record/MemoryLogBufferBuilder.java
@@ -98,7 +98,7 @@ public class MemoryLogBufferBuilder {
     private final long logAppendTime;
     private final long pid;
     private final short epoch;
-    private final int baseSequence;
+    private final long baseSequence;
     private final int writeLimit;
 
     private MemoryLogBuffer builtLogBuffer;
@@ -117,7 +117,7 @@ public class MemoryLogBufferBuilder {
                                   long logAppendTime,
                                   long pid,
                                   short epoch,
-                                  int baseSequence,
+                                  long baseSequence,
                                   int writeLimit) {
         this.magic = magic;
         this.timestampType = timestampType;
